[CSL-1279] Rename Pos.Binary.GodTossing.*
