+++ lib/src/Pos/Binary.hs
@@ -9,7 +9,7 @@ import           Pos.Binary.Class
 import           Pos.Binary.Communication ()
 import           Pos.Binary.Core          ()
 import           Pos.Binary.Crypto        ()
-import           Pos.Binary.GodTossing    ()
+import           Pos.Binary.Ssc           ()
 import           Pos.Binary.Merkle        ()
 import           Pos.Binary.Relay         ()
 import           Pos.Binary.Ssc           ()
