--- lib/src/Pos/Binary/Ssc.hs
@@ -1,6 +1,0 @@
--- | GodTossing serialization instances
-
-module Pos.Binary.Ssc () where
-
-import           Pos.Binary.GodTossing ()
-
