+++ lib/test/Test/Pos/Cbor/CborSpec.hs
@@ -35,7 +35,7 @@ import           Pos.Binary.Communication          ()
 import           Pos.Binary.Core.Fee               ()
 import           Pos.Binary.Core.Script            ()
 import           Pos.Binary.Crypto                 ()
-import           Pos.Binary.GodTossing             ()
+import           Pos.Binary.Ssc                    ()
 import           Pos.Binary.Infra                  ()
 import           Pos.Binary.Relay                  ()
 import           Pos.Core.Types                    (ScriptVersion)
