+++ ssc/Pos/Binary/Ssc/Core.hs
@@ -1,6 +1,6 @@
 -- | Serialization of core types from GodTossing SSC.
 
-module Pos.Binary.GodTossing.Core
+module Pos.Binary.Ssc.Core
        (
        ) where
