+++ ssc/Pos/Ssc/GState.hs
@@ -24,7 +24,7 @@ import           Formatting                       (build, sformat, (%))
 import           System.Wlog                      (WithLogger, logDebug, logInfo)
 import           Universum
 
-import           Pos.Binary.GodTossing            ()
+import           Pos.Binary.Ssc                   ()
 import           Pos.Core                         (EpochIndex (..),
                                                    HasConfiguration,
                                                    SharedSeed, SlotId (..),
